# services/api/openapi.yaml
...
openapi: 3.0.3
info:
  title: Báscula API
  version: "1.0.0"
  description: |
    Multi-tenant farm labour and payroll API. One binary, one Postgres, four
    roles. Isolation is row level security keyed on `app.farm_id`, set from the
    access token inside the transaction that serves the request.
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /v1/prices/weeks/{monday}:
    parameters:
      - name: monday
        in: path
        required: true
        description: |
          The week is named by its Monday's ISO date. Any other day is a 400 —
          the phone's old "2026-W33" form is obsolete.
        schema: { type: string, format: date, example: "2026-08-24" }
    get:
      tags: [money]
      operationId: getWeekPrice
      summary: What a kilo is worth this week
      description: |
        The week's override if the owner set one, otherwise the farm's standing
        price. A work record priced by the week reads this at settlement time,
        not at write time.
      x-action: prices.read
      x-roles: [owner, admin]
      responses:
        "200":
          description: The price in force for that week.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WeekPrice" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [money]
      operationId: setWeekPrice
      summary: Set this week's price
      description: |
        Owner only. Setting a past week's price reprices every weighing of that
        week that has not been settled yet, and none that has: a settled line
        keeps the price it was settled at.
      x-action: prices.write
      x-roles: [owner]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [priceCents]
              properties:
                priceCents: { type: integer, format: int64, minimum: 1 }
      responses:
        "200":
          description: Stored.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WeekPrice" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }

  # ---------------------------------------------------------------------------
  # Money
  # ---------------------------------------------------------------------------
  /v1/pending:
    get:
      tags: [money]
      operationId: pending
      summary: What a worker is owed for but has not been paid
      description: |
        Payables in the range that no live settlement has claimed, each already
        priced. Two freezing moments meet here: a line priced by the week is
        priced now, at read time; every other line was frozen when it was
        written and is only read back.

        Deliberately not filtered by pay scheme. Filtering it would bring back
        the two-path problem that the single payable table removed — a picker
        who also cleared brush must receive one settlement.

        The worker is confirmed to be ours before anything is summed.
      x-action: pending.read
      x-roles: [owner, admin]
      parameters:
        - name: workerId
          in: query
          required: true
          schema: { type: string, format: uuid }
        - name: from
          in: query
          required: true
          schema: { type: string, format: date }
        - name: to
          in: query
          required: true
          schema: { type: string, format: date }
      responses:
        "200":
          description: The unsettled lines and their total.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PendingResult" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/balances:
    get:
      tags: [money]
      operationId: listBalances
      summary: Every worker's position
      description: Derived from the ledger, largest credit first.
      x-action: balances.read
      x-roles: [owner, admin]
      responses:
        "200":
          description: One row per worker.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Balance" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
  /v1/settlements/{id}:
    parameters:
      - $ref: "#/components/parameters/PathID"
    get:
      tags: [money]
      operationId: getSettlement
      summary: One settlement with its lines
      x-action: settlements.read
      x-roles: [owner, admin]
      responses:
        "200":
          description: The settlement.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Settlement" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/settlements/{id}/void:
    parameters:
      - $ref: "#/components/parameters/PathID"
    post:
      tags: [money]
      operationId: voidSettlement
      summary: Cancel a settlement
      description: |
        Without editing a thing. The lines keep their rows for the record but
        gain a `voidedAt`, which is what releases their payables from the lock;
        the earning is cancelled by a reversal, never deleted.
      x-action: settlements.void
      x-roles: [owner, admin]
      responses:
        "200":
          description: The settlement, now void.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Settlement" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: SETTLEMENT_ALREADY_VOID.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /v1/ledger/{id}/reverse:
    parameters:
      - $ref: "#/components/parameters/PathID"
    post:
      tags: [money]
      operationId: reverseLedgerEntry
      summary: Undo a movement
      description: |
        The only way back. Nothing in the ledger is edited or deleted — the
        database has triggers that forbid both — so a mistake is cancelled by a
        movement that is its exact opposite, once. A partial unique index on
        `reversesId` is what makes it once; a second attempt is 409
        ALREADY_REVERSED. A reversal cannot itself be reversed.
      x-action: ledger.reverse
      x-roles: [owner, admin]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                note: { type: string, nullable: true }
      responses:
        "201":
          description: The reversing movement.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LedgerEntry" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: ALREADY_REVERSED.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        A 15-minute access token with claims `sub`, `farm_id`, `role`,
        `device_id`. The farm comes from here and from nowhere else.
  parameters:
    PathID:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    BadRequest:
      description: BAD_REQUEST — malformed body, unknown field, or a bad value.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: UNAUTHORIZED — no token, or a token this server did not issue.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: FORBIDDEN — the permission table refused this role.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: |
        NOT_FOUND. Also what a resource of another farm answers, and what an id
        that is not a UUID answers: nothing with that id exists here.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    ErrorCode:
      type: string
      description: |
        The machine-readable half of every failure. The client branches on
        this; the translation lives in the client, which is why `message` is
        English.

        This enum is compared against domain.AllCodes() by the contract test,
        so a code cannot exist in the server without appearing here.
      enum:
        - BAD_REQUEST
        - UNAUTHORIZED
        - FORBIDDEN
        - NOT_FOUND
        - CONFLICT
        - INTERNAL
        - TENANT_NOT_SET
        - INVALID_CREDENTIALS
        - EMAIL_NOT_VERIFIED
        - EMAIL_TAKEN
        - TOKEN_EXPIRED
        - TOKEN_REUSED
        - RATE_LIMITED
        - FARM_LIMIT_REACHED
        - FARM_SUSPENDED
        - WORK_RECORD_SETTLED
        - PAYABLE_ALREADY_CLAIMED
        - SETTLEMENT_ALREADY_VOID
        - ALREADY_REVERSED
        - NOTHING_TO_SETTLE
        - AMOUNT_EXCEEDS_BALANCE
        - INVALID_GEOMETRY
        - PLOT_HAS_ACTIVE_CROPS
        - NO_RATE_IN_FORCE
        - RANGE_NEEDS_FROZEN_RATE
        - DUPLICATE_DOCUMENT
        - DUPLICATE_NAME
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { $ref: "#/components/schemas/ErrorCode" }
            message:
              type: string
              description: English, for a developer and a log. Never shown to a picker.
            details:
              type: object
              additionalProperties: true
              description: |
                What a client needs to recover. The canonical case is
                PAYABLE_ALREADY_CLAIMED, where `winningSettlement` lets the
                phone re-derive its state.
...
    PayScheme:
      type: string
      description: |
        How an activity pays. Spanish values because they are the values stored
        in the Postgres enum, which the design documents fixed.
      enum: [contrato, tiempo, unidad_trabajo]
    RateSource:
      type: string
      description: |
        When the price froze. `explicit` — the caller named it, frozen at write
        time. `activity_dated` — the activity's rate in force on the record's
        day, frozen at write time. `weekly_price` — read from the week at
        settlement time, which is what the phone does today. The two derived
        forms require a single-day record.
      enum: [explicit, activity_dated, weekly_price]
    LedgerKind:
      type: string
      description: |
        These six describe a coffee pickup, a day's wage and a pruning contract
        equally well, which is why generalising the payable side required no
        change here at all.
      enum: [devengo, pago, anticipo, deduccion, ajuste, reverso]
    PayMethod:
      type: string
      enum: [efectivo, transferencia, otro]
...
    Balance:
      type: object
      description: |
        Derived from the ledger, never stored: a stored total is a total that
        one day lies. Positive means the farm owes them.
      required: [workerId, earnedCents, paidCents, deductedCents, balanceCents]
      properties:
        workerId: { type: string, format: uuid }
        earnedCents: { type: integer, format: int64 }
        paidCents: { type: integer, format: int64 }
        deductedCents: { type: integer, format: int64 }
        balanceCents: { type: integer, format: int64 }
        lastMovementOn: { type: string, format: date-time, nullable: true }
    LedgerEntry:
      type: object
      description: |
        One movement. Append-only: nothing here is ever edited or deleted, and
        the database has triggers that forbid both. Signs are part of the
        contract — `pago`, `anticipo` and `deduccion` are negative, `devengo`
        is positive, and a `reverso` carries the exact opposite of what it
        cancels.
      required: [id, workerId, kind, amountCents, date, createdAt]
      properties:
        id: { type: string, format: uuid }
        workerId: { type: string, format: uuid }
        kind: { $ref: "#/components/schemas/LedgerKind" }
        amountCents: { type: integer, format: int64 }
        date: { type: string, format: date-time }
        settlementId: { type: string, format: uuid, nullable: true }
        method:
          allOf: [{ $ref: "#/components/schemas/PayMethod" }]
          nullable: true
        note: { type: string, nullable: true }
        reversesId: { type: string, format: uuid, nullable: true }
        createdAt: { type: string, format: date-time }
...
    Payable:
      type: object
      description: One unsettled work record, priced.
      required: [payableId, activityId, activity, payScheme, rateSource, quantity, rateCents, amountCents]
      properties:
        payableId: { type: string, format: uuid, description: The work record's id. }
        activityId: { type: string, format: uuid }
        activity: { type: string }
        payScheme: { $ref: "#/components/schemas/PayScheme" }
        rateSource: { $ref: "#/components/schemas/RateSource" }
        quantity: { type: string }
        unitId: { type: string, format: uuid, nullable: true }
        date: { type: string, format: date-time }
        weekStart: { type: string, format: date-time }
        rateCents: { type: integer, format: int64 }
        amountCents: { type: integer, format: int64 }
        voided: { type: boolean, description: Only meaningful inside a settlement. }
    PendingResult:
      type: object
      required: [workerId, from, to, items, totalCents]
      properties:
        workerId: { type: string, format: uuid }
        from: { type: string, format: date-time }
        to: { type: string, format: date-time }
        items:
          type: array
          items: { $ref: "#/components/schemas/Payable" }
        totalCents: { type: integer, format: int64 }
...
    Settlement:
      type: object
      required: [id, workerId, periodStart, periodEnd, grossCents, status, items]
      properties:
        id: { type: string, format: uuid }
        workerId: { type: string, format: uuid }
        periodStart:
          type: string
          format: date-time
          description: |
            The period actually covered — the Monday of the earliest payable
            taken in — not the window the caller asked over.
        periodEnd: { type: string, format: date-time }
        grossCents: { type: integer, format: int64 }
        status: { type: string, enum: [open, void] }
        note: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        voidedAt: { type: string, format: date-time, nullable: true }
        items:
          type: array
          items: { $ref: "#/components/schemas/Payable" }
    WeekPrice:
      type: object
      required: [weekStart, priceCents]
      properties:
        weekStart: { type: string, format: date, description: An ISO Monday. }
        priceCents: { type: integer, format: int64 }